adding primary gameplay loop script
this is the main script that controls the primary gameplay loop. it use to utilize a heavily edited version of v_yriss's disaster kit v2, but is now being rewritten from the ground up. having to learn a lot to get the majority of the functionality working, and there are many things that don't work currently. at the moment, the timer will run out and no hazards will spawn, which breaks the remainder of the loop. i will continue to tinker with this and update it when applicable.