Update Main Gameplay Loop.lua
Reworked the gear handling from scratch to work with a folder in ServerStorage rather than a list of Asset IDs. Made it so the function pulls gear from said folder, puts them into a table, verifies that all 6 table entries are unique, then copies those entries and gives them to the player.